+++ packages/core/src/recaptcha/ReCaptcha.tsx
@@ -9,6 +9,8 @@ export interface ReCaptchaProps {
   tabIndex?: number
   value?: string
   onChange: (token: string | undefined) => void
+  onErrored?: () => void
+  asyncScriptOnLoad?: () => void
 }
 
 const width = 304
@@ -35,6 +37,14 @@ export class ReCaptcha extends PureComponent<ReCaptchaProps> {
 
   private onChange = (token: string | null) => {
     this.props.onChange(token === null ? undefined : token)
+  }
+
+  private onErrored = () => {
+    this.props.onErrored?.()
+  }
+
+  private asyncScriptOnLoad = () => {
+    this.props.asyncScriptOnLoad?.()
   }
 
   private resize = () => {
@@ -92,10 +102,11 @@ export class ReCaptcha extends PureComponent<ReCaptchaProps> {
             size="normal"
             type="image"
             onChange={this.onChange}
+            onErrored={this.onErrored}
+            asyncScriptOnLoad={this.asyncScriptOnLoad}
           />
         </Pos>
       </Pos>
     )
   }
-
 }